requirements.txt
streamlit==1.22.0
tensorflow==2.19.0
numpy==1.26.4
Pillow==10.4.0
matplotlib==3.9.2
gdown==5.2.0